selling/buying bonds

+++ src/com/etc/Valbz.java
@@ -1,6 +1,6 @@
 package com.etc;
 
 public class Valbz extends Security {
-    private Integer fairValue;
+    private Float fairValue;
 
 }
